requirements.txt
fastapi
pydantic
uvicorn
python-decouple
peewee
PyMySQL
email-validator
pre-commit
black
flake8
pytz
pyjwt
passlib
redis[hiredis]
langchain
requests
langchain-openai
ctransformers
langchain-core
langchain-community
huggingface_hub
transformers
langchainhub